1.16-pre1

+++ minecraft/src/net/minecraft/block/FallingBlock.java
@@ -68,9 +68,9 @@ public class FallingBlock extends Block {
 		if (random.nextInt(16) == 0) {
 			BlockPos blockPos = pos.down();
 			if (canFallThrough(world.getBlockState(blockPos))) {
-				double d = (double)pos.getX() + (double)random.nextFloat();
+				double d = (double)pos.getX() + random.nextDouble();
 				double e = (double)pos.getY() - 0.05;
-				double f = (double)pos.getZ() + (double)random.nextFloat();
+				double f = (double)pos.getZ() + random.nextDouble();
 				world.addParticle(new BlockStateParticleEffect(ParticleTypes.FALLING_DUST, state), d, e, f, 0.0, 0.0, 0.0);
 			}
 		}
