+++ minecraft/src/net/minecraft/client/sound/EntityTrackingSoundInstance.java
@@ -19,9 +19,9 @@ public class EntityTrackingSoundInstance extends MovingSoundInstance {
 		this.volume = volume;
 		this.pitch = pitch;
 		this.entity = entity;
-		this.x = (float)this.entity.getX();
-		this.y = (float)this.entity.getY();
-		this.z = (float)this.entity.getZ();
+		this.x = (double)((float)this.entity.getX());
+		this.y = (double)((float)this.entity.getY());
+		this.z = (double)((float)this.entity.getZ());
 	}
 
 	@Override
@@ -34,9 +34,9 @@ public class EntityTrackingSoundInstance extends MovingSoundInstance {
 		if (this.entity.removed) {
 			this.setDone();
 		} else {
-			this.x = (float)this.entity.getX();
-			this.y = (float)this.entity.getY();
-			this.z = (float)this.entity.getZ();
+			this.x = (double)((float)this.entity.getX());
+			this.y = (double)((float)this.entity.getY());
+			this.z = (double)((float)this.entity.getZ());
 		}
 	}
 }
