+++ minecraft/src/net/minecraft/client/sound/SoundInstance.java
@@ -27,11 +27,11 @@ public interface SoundInstance {
 
 	float getPitch();
 
-	float getX();
+	double getX();
 
-	float getY();
+	double getY();
 
-	float getZ();
+	double getZ();
 
 	SoundInstance.AttenuationType getAttenuationType();
